requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
cairosvg>=2.7.1
pyautogui>=0.9.54
pyperclip>=1.9.0
numpy>=1.26.0
opencv-python>=4.10.0
mss>=10.0.0
PyAudio>=0.2.14
SpeechRecognition>=3.10.4
openwakeword>=0.6.0
faster-whisper>=1.0.3
send2trash>=1.8.2
playwright>=1.48.0
psutil>=5.9.8
beautifulsoup4>=4.12.3
youtube-transcript-api>=0.6.2
yt-dlp>=2025.1.15
ddgs>=8.0.0
duckduckgo-search>=6.3.2
